drf_api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.3.2
distro==1.9.0
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-grappelli==4.0.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
multidict==6.4.4
openai==0.28.0
packaging==24.2
pillow==11.2.1
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
sniffio==1.3.1
sqlparse==0.5.3
stripe==11.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.20.0
